+++ docs/CHANGELOG.md
@@ -11,6 +11,7 @@ Released on: ???
 - Added Dwarven Lanterns to the deployables tracker and alerts (disabled by default).
 - Added Rain/Thunderstorm timer overlay for Birch Park/Spider Den, and alert when Rain/Thunderstorm ends soon (disabled by default).
 - Added settings to customize compact catch messages written to your chat when you catch any sea creature. You can type your own text for double hook and catch message.
+- Added setting to hide other players near your bobber when a fishing rod is casted (disabled by default).
 
 ## Bugfixes